feat(CellValue): makes CellValue an interface
